pyproject.toml
[build-system]
requires = [
    "scikit_build_core",
    "pybind11",
    "cmake>=3.28",
    "ninja",
    "numpy",
]
build-backend = "scikit_build_core.build"

[project]
name = "alphafold3"
version = "3.0.1"
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "absl-py",
    "chex",
    "dm-haiku==0.0.13",
    "dm-tree",
    "jax==0.4.34",
    "jax[cuda12]==0.4.34",
    "jax-triton==0.2.0",
    "jaxtyping==0.2.34",
    "numpy",
    "rdkit==2024.3.5",
    "triton==3.1.0",
    "tqdm",
    "typeguard==2.13.3",
    "zstandard",
]

[project.optional-dependencies]
test = ["pytest>=6.0"]

[tool.scikit-build]
wheel.exclude = [
    "**.pyx",
    "**/CMakeLists.txt",
    "**.cc",
    "**.h"
]
sdist.include = [
    "LICENSE",
    "OUTPUT_TERMS_OF_USE.md",
    "WEIGHTS_PROHIBITED_USE_POLICY.md",
    "WEIGHTS_TERMS_OF_USE.md",
]

[tool.cibuildwheel]
build = "cp3*-manylinux_x86_64"
manylinux-x86_64-image = "manylinux_2_28"

[project.scripts]
build_data = "alphafold3.build_data:build_data"
